extracted CaptureTheFlagUtils and overhauled tests
